+++ package.json
@@ -9,8 +9,8 @@
     "test": "npx jest --config jest.config.js --verbose",
     "build": "./build.sh",
     "build:api": "cd api && npx webpack",
-    "build:core": "cd utils && npx webpack",
-    "build:utils": "cd core && npx webpack",
+    "build:core": "cd core && npx webpack",
+    "build:utils": "cd utils && npx webpack",
     "serve:api": "npx nodemon --exec node ./api/server.js",
     "serve:api-dist": "node ./api/dist/server.bundle.cjs",
     "serve:core": "cd ./core && npx nodemon --exec npx webpack --watch",
