+++ utils/index.js
@@ -1,12 +1,17 @@
 import isNonNegativeNumber from "./src/isNonNegativeNumber.js";
 import isEmpty from "./src/isEmpty.js";
 import formatToDecimalPoint from "./src/formatToDecimal.js";
+import validAmmortizationPeriod from './src/validateAmmortizationPeriod.js';
+import validatePaymentSchedule from "./src/validatePaymentSchedule.js";
+
 
 
 const utilities = {
     isNonNegativeNumber,
     isEmpty,
-    formatToDecimalPoint
+    formatToDecimalPoint,
+    validAmmortizationPeriod,
+    validatePaymentSchedule
 };
 
 export default utilities;
